==> 1_EXTRACTION.ipynb <==
### 1. Extraction

##### Tools

import requests as req
import pandas as pd
from selenium import webdriver
from webdriver_manager.chrome import ChromeDriverManager
import warnings   # para quitar warnings
from chromedriver_py import binary_path # this will get you the path variable
import re

### 1st Scraping method : Selenium 

In order to obtain the information from a specific 'URL', will use variouse libraries and modules, that will be specify in each of the 5 steps.

### Table_1: Population in india by region trough the years (from 1951 to 2011)

**Data From The Reserve Bank of India (RBI).** 

Is India's central bank and regulatory body and is responsible for the issue and supply of the Indian rupee and the regulation of the Indian banking system. 

**Step 1**: Define the PATH using 'ChromeDriverManager' module form webdriver_manager library, that will be used  to perform the scarpe trough Google Chrome app.

PATH=ChromeDriverManager().install() #PATH

**Step 2**: Define the url will like to scrap.

url = 'https://m.rbi.org.in/scripts/PublicationsView.aspx?id=20660'

**Step 3:** Difine the 'driver' with the module, 'webdriver' from **Selenium**, Module 'Chrome' and the PATH we defined before. The driver, is the boot that will acces the serch einge and look for the specified web. 


driver = webdriver.Chrome(PATH) #Driver
driver.get(url)

**Step 4:** Using the 'driver' + 'find_element_by_xpath' and specifying the 'XPATH' from the html element that will like to obtain also adding the '.text' module, will obtain directly the aimed information, without html tags. 

Table = driver.find_element_by_xpath('//*[@id="doublescroll"]/table[2]/tbody/tr[4]/td/table').text

**Step 5:** Finish the process with 'driver' + 'quit()' module.

driver.quit()


**Data Cleaning**

Will perform some data cleaning methods, in order to create a proper DataFrame:

**Extracting the India States names:**

* **States_0:** First will split the text (type string), by the 'line break'.
* **States_1:** In a comprehension list, will use regex method sub(substitude) for a whitespace.
* **States_2:** In a comprehension list, we eraised the right whitespaces, left from the regex function. 
* **States_2[1:-1]:** In order to exclude the column names and bottom text, we select by list index, the content will like to keep. 


States_0 = Table.split('\n') 

States_1 = [''.join(re.sub(r'[^a-zA-Z]',' ' , i)) for i in States_0[1:38]]

States_2 = [s.rstrip() for s in States_1] 

States_2[1:-1]

**Extracting the India Population by Region:**

* **values:** In a comprehension list, will use regex method to find all numbers and exclude the slash '/' character.
* **values_1:** In a comprehension list, will use a build in python method to replace the En Dash '-' character with a 0.
* **values_2:** In a comprehension list, we eraised the left whitespaces. 
* **values_3:** In a comprehension list, we substitude the spaces between the numbers with comas.
* **values_4:** This was a thoght one! in order to remove the single quotes, to be able to amnage the data as list, i used the 'map' function to unpack the string, convert them into integers and split by the ',' to obatin the different lists.
* **values_5:** Since i had a list of list, i couldn't convert it into a DataFrame yet, so i convert it into a tupple, applyied a function to replace the Tuple brackets, for nothing. It worked! 
* **values_6:** I copied the result of the function return, since it didn't lend me used from it.

values = [''.join(re.findall(r'[^a-zA-Z\/]', i)) for i in States_0[1:38]]

values_1 = [d.replace('-','0')for d in values]

values_2 = [r.lstrip() for r in values_1]

values_3 = [re.sub(' ',',',c) for c in values_2]
# Substitude the spaces between the numbers with comas

values_4 = [list(map(int,values_3[c].split(','))) for c in range(len(values_3))]
# INCREDIBLYYYYY!!! we remove the single quotes of each list!!!


values_4[1:]

values_5 = tuple(values_4[1:])

def db(list1):
    return str(list1).replace('(','').replace(')','')

Test_1 = db(values_5) # porque me permite copiar el data set (directamente de aqui), pero si lo copio de una lista de lista no me lo permite??

values_6 = [31, 64, 115, 189, 281, 356, 381], [31115, 35983, 43503, 53551, 66508, 76210, 84581], [0, 337, 468, 632, 865, 1098, 1384], [8029, 10837, 14625, 18041, 22414, 26656, 31206], [29085, 34841, 42126, 52303, 64531, 82999, 104099], [24, 120, 257, 452, 642, 901, 1055], [7457, 9154, 11637, 14010, 17615, 20834, 25545], [42, 58, 74, 104, 138, 220, 344], [49, 37, 63, 79, 102, 158, 243], [1744, 2659, 4066, 6220, 9421, 13851, 16788], [547, 590, 795, 1008, 1170, 1348, 1459], [16263, 20633, 26697, 34086, 41310, 50671, 60440], [5674, 7591, 10036, 12922, 16464, 21145, 25351], [2386, 2812, 3460, 4281, 5171, 6078, 6865], [3254, 3561, 4617, 5987, 7837, 10144, 12541], [9697, 11606, 14227, 17612, 21844, 26946, 32988], [19402, 23587, 29299, 37136, 44977, 52851, 61095], [13549, 16904, 21347, 25454, 29099, 31841, 33406], [21, 24, 32, 40, 52, 61, 64], [18615, 23218, 30017, 38169, 48566, 60348, 72627], [32003, 39554, 50412, 62783, 78937, 96879, 112374], [578, 780, 1073, 1421, 1837, 2294, 2856], [606, 769, 1012, 1336, 1775, 2319, 2967], [196, 266, 332, 494, 690, 889, 1097], [213, 369, 516, 775, 1210, 1990, 1979], [14646, 17549, 21945, 26370, 31660, 36805, 41974], [317, 369, 472, 604, 808, 974, 1248], [9161, 11135, 13551, 16789, 20282, 24359, 27743], [15971, 20156, 25766, 34262, 44006, 56507, 68548], [138, 162, 210, 316, 406, 541, 611], [30119, 33687, 41199, 48408, 55859, 62406, 72147], [639, 1142, 1556, 2053, 2757, 3199, 3674], [60274, 70144, 83849, 105137, 132062, 166198, 199812], [2946, 3611, 4493, 5726, 7051, 8489, 10086], [26300, 34926, 44312, 54581, 68078, 80176, 91276], [361088, 439235, 548160, 683329, 846421, 1028737, 1210855]

**Creating the Data Frames with Pandas.**

I created tow different Data Frames, one for the names of Indian Sates (df_c)and another for the population values (df_n):

* df_n_1: Toght one also! But applying the 'pd.Series' pandas Method i was able to split the list in different columns of the data Frame.

df_c = pd.DataFrame({'States':States_2[1:-1]})

df_n = pd.DataFrame({'years': values_6}) 

df_n_1 = df_n.years.apply(pd.Series)

df_n_1.columns = ['1951', '1961','1971', '1981', '1991', '2001', '2011']

df_n_1.sample()

**Exporting the data in 'CSV' Format and specifying to not build a new index,and saveing it into the data folder.**

df_n_1.to_csv('data/DF_India_Population_Extrac.csv',index=False)

df_c.to_csv('data/DF_India_States.csv',index=False)

### 2nd Scraping Method:  Panda's Library 

* **Table_2:** NSDP(Net state domestic product) per capita, trhoug the years by India states.

* **Table_3:** Covid cases, by India states.

### Table_2 : India NSDP per capita by Region trough the years (from 1996-2011).

**Data from Wikipedia**

The Free Encyclopedia

**Extacting and Data Cleaning, Table_2.**

Will perform some data cleaning methods, in order to create a proper DataFrame:

* **Step_0:** Incredibly, There is a method in Panda's library called 'read_html' that giving the url, will read the html code, find 'table' tags and export them as DataFrames. If there are several 'tables' in the url, you will be able to acessed them, by an index list type. [0] will be the fist table, and so on. There is only one Step, The Step!
* **Step_1:** I prepare the data frame, to be able to work with it. first step, will be to crush the 'instance tribute', replace all white space for and underscore '_' and all lotters in lower case..
* **Step_2:** Renaiming the columns names.
* **Step_3:** Droping 'Rank' column. 
* **Step_4:** The numbers are separated by ',' this will create a conflict while working with the data. Using comprehension, list, iloc, lambda and replace, were are able to acces all data from our dataframe and clean it up.
* **Step_5:** Using the same method but to eliminate the Indian Rupee '₹' character.




**Table_2: Step 0**

data_ec = pd.read_html('https://en.wikipedia.org/wiki/List_of_Indian_states_and_union_territories_by_GDP_per_capita',skiprows=0)[4]

data_copy = data_ec.copy()

df_economy = pd.DataFrame(data_ec)

df_economy.describe()

df_economy.columns

**Table_2: Step_1**

lista_col = [col.replace(' ','_').lower() for col in df_economy.columns]

df_economy.columns = lista_col

**NSDP ( Net state domestic product) PER CAPITA INDIA.**
* ***Indian Rupee***, conversion to Euros, 1€  = 85.468 ₹. or 1₹ =  0.0116941€

df_economy.sample()

**Table_2: Step_2**

df_economy.columns

df_economy.rename(columns={'state/union_territory':'State','nsdp_per_capita_(nominal)1996–97[4]':'1996_to_1997','nsdp_per_capita_(nominal)2000–01[4]':'2000_to_2001','nsdp_per_capita_(nominal)2005–06[4]':'2005_to_2006','nsdp_per_capita_(nominal)2010–11[4]':'2010_to_2011','growth_from_2000–2010':'Growth_from_2000_to_2010'},inplace=True)


**Table_2: Step_3**

df_economy.drop(['rank'],axis=1,inplace=True)

df_economy.sample()

df_economy.iloc[:,1:5].sample()

**Table_2: Step_4**

df_economy.iloc[:,1:5] = df_economy.iloc[:,1:5].apply(lambda x: x.str.replace(',', ''), axis=1)

#df.replace(r'[()]', np.nan, regex=True,
          #inplace=True)    

**Table_2: Step_5**

df_economy.iloc[:,1:5] = df_economy.iloc[:,1:5].apply(lambda x: x.str.replace('₹', ''), axis=1)


**Convert indian rupee into Euros.**

* df_economy.select_dtypes(exclude=['State', 'Growth_from_2000_to_2010']) * 0.0116941 


* link it to this web: https://www.xe.com/currencycharts/?from=INR&to=EUR

df_economy.sample(2)

**Table_2: Step_5**

df_economy.to_csv('data/DF_India_nsdp_per_capita_Extract.csv', index=False)

### Table_3: Covid cases, fatal and no fatal in by Regions in India.

**Data From: PRS Legislative Research** 

Commonly referred to as PRS, is an Indian non-profit organisation that was established in September 2005 as an independent research institute to make the Indian legislative process better informed, more transparent and participatory. PRS is based in New Delhi.

**Extacting and Data Cleaning, Table_3.**

Will perform some data cleaning methods, in order to create a proper DataFrame:

* **Step_0:** Extracting data from PRS.
* **Step_1:** Rename columns.
* **Step_2:** Delete non value column.
* **Step_3:** Reset index.
* **Step_4:** Exporting data.

**Table_3: Step_0**

data_covid = pd.read_html('https://prsindia.org/covid-19/cases',skiprows=0)[0]

df_covid = pd.DataFrame(data_covid)

df_covid.describe()

df_covid.sample()

**Table_3: Step_1**

df_covid.rename(columns={'State/UT':'State','Cured/Discharged':'Cured'},inplace=True)

**Table_3: Step_2**

df_covid.drop(['#'],axis=1,inplace=True)

**Table_3: Step_3**

df_covid_19 = df_covid.reset_index(drop=True)

df_covid_19.sample()

**Table_3: Step_4**

df_covid_19.to_csv('data/DF_India_Covid_Incidents_Extract.csv',index=False)
