==> 2_TRANSFORMATION.ipynb <==
## 2_TRANSFORMATION.

import pandas as pd

#### 1. Table_1 POPULATION EVOLUTION INDIA

df_n_1 = pd.read_csv('data/DF_India_Population_Extrac.csv',encoding = "ISO-8859-1")

df_c = pd.read_csv('data/DF_India_States.csv',encoding = "ISO-8859-1")

df_c.sample()

df_n_1.sample()

df_popul = pd.concat([df_c, df_n_1], axis=1, join='inner')

df_population = pd.DataFrame(df_popul)

df_population.sample(3)

* I sorted the data by State names, in ascending order. Although it wasn't strictly nessesary, since the Pandas 'merge()' method, will automatically, match the value names that are equal. Even though sortering the data was useful to have a clear view of the merging process. 

df_population.sort_values(by="States",ascending=True, inplace=True)


df_population.reset_index(drop=True).sample(2)

df_population.sample(3)

df_population.shape

df_population.to_csv('data/DF_India_Population.csv',index=False)

#### 2. Table_2 INDIA EVOLUTION ECONOMY

df_economy = pd.read_csv("data/DF_India_nsdp_per_capita_Extract.csv",encoding = "ISO-8859-1")


df_economy.sort_values(by="State",ascending=True, inplace=True)

df_economy.reset_index(drop=True).head(1)

df_economy.head(1)

df_economy.shape

df_economy.to_csv('data/DF_India_Economy.csv',index=False)

#####  MERGING_1, POPULATION EVOLUTION INDIA &  NSD PER CAPITA INDIA.

MERGE_1 = df_population.merge(df_economy, left_on='States', right_on='State', suffixes=['_izq', 'dcha'])



MERGE_1.head(3)

#### 3. Table_3 COVID INCIDENTS BY REGION INDIA

df_covid = pd.read_csv("data/DF_India_Covid_Incidents_Extract.csv",encoding = "ISO-8859-1")

df_covid.sort_values(by="State",ascending=True, inplace=True)

df_covid.sample()

df_covid.shape

df_covid.to_csv('data/DF_India_Covid.csv',index=False)

#### MERGING_2  POPULATION EVOLUTION INDIA &  NSD PER CAPITA INDIA & COVID INCIDENTS BY REGION INDIA

MERGE_2 = MERGE_1.merge(df_covid, left_on='States', right_on='State', suffixes=['_izq', 'dcha'])

MERGE_2.head(3)
