==> 3_LOAD.ipynb <==
## 3_LOAD

##### Tools

import mysql.connector as mysql # Stablish connection to MySQL

from mysql.connector import MySQLConnection # Using class for cunction 

import mysql.connector as conn # Calling config files???

#from python_to_myql_config import read_db_config

import pandas as pd

import csv

cnx = conn.connect(host='localhost',
                  user='root',
                  password='admin')
                  #database= 'india')
#stablishing connection



mycursor = cnx.cursor() # Cursor, allows us to do queries,send/retrive data




**Delette Data Base:**

#mycursor.execute('DROP DATABASE india')

**3.1 CREO NUEVA BASE DE DATOS, LLAMADA india**  

**Create Data Base:**

#mycursor.execute('CREATE DATABASE india')

**Stablishing connection:**


cnx = conn.connect(host='localhost',
                  user='root',
                  password='admin',
                  database= 'india')
#stablishing connection



mycursor = cnx.cursor() # Cursor, allows us to do queries,send/retrive data



### 1. India Population Growth by Region from 1951 to 2011:

#### 1.1 Import DF_India_Population, look to it's data.

df_popul=pd.read_csv('data/DF_India_Population.csv')

#### See data type of the column, first!

df_popul.dtypes

### ELIMANTE TABLE!!!
sql = "DROP TABLE India_population"
mycursor.execute(sql)
### ELIMANTE TABLE!!!

#### 1.2 Create table, India_population: 
* Assing the PK (Primary Key) to the column "States_India_ID". 
* Create the columns, 
* Specifying their value types .

sql_1 = """ CREATE TABLE India_population(States_India_ID VARCHAR(50) NOT NULL, 
year_1951 INT NOT NULL, year_1961 INT NOT NULL, year_1971 INT NOT NULL,
year_1981 INT NOT NULL, year_1991 INT NOT NULL, year_2001 INT NOT NULL, year_2011 INT NOT NULL, PRIMARY KEY (States_India_ID))

"""

mycursor.execute(sql_1)

mycursor.execute('select * from India_population')
mycursor.fetchall()

#### 1.3 Insert the data to the  India_population

df_popul=pd.read_csv('data/DF_India_Population.csv')

mycursor.executemany("""
INSERT INTO India_population(States_India_ID, 
year_1951, year_1961, year_1971,year_1981, 
year_1991, year_2001, year_2011) VALUES(%s,%s,%s,%s,%s,%s,%s,%s)
""", list(df_popul.values))

mycursor.execute('select * from India_population ')
len(mycursor.fetchall())

#### 1.4 Commit the changes to be updated in the MySQL Server.

cnx.commit()

### 2. India Net State Domestic Product Per Capita , by Region. (Nominal?-What does it means?).

#### 2.1 Import DF_India_Economy, look to it's data.

df_economy=pd.read_csv('data/DF_India_Economy.csv')

df_economy.dtypes

#### 2.2 Create Table India_NSDP_Evolution:

* Assing the PK (Primary Key) to the column "States_India_ID". 
* Create the columns, 
* Specifying their value types .

sql_2 = """ CREATE TABLE India_NSDP_Evolution(
States_India_ID VARCHAR(50) NOT NULL, 
1996_to_1997 INT NOT NULL, 
2000_to_2001 INT NOT NULL, 
2005_to_2006 INT NOT NULL,
2010_to_2011 INT NOT NULL,
Growth_from_2000_to_2010 VARCHAR(50) NOT NULL, 
PRIMARY KEY (States_India_ID))

"""

mycursor.execute(sql_2)

mycursor.execute('select * from India_NSDP_Evolution')
mycursor.fetchall()

#### 2.3 Insert the data to the  Table India_NSDP_Evolution:

df_Economy=pd.read_csv('data/DF_India_Economy.csv')

mycursor.executemany("""
INSERT INTO India_NSDP_Evolution(
States_India_ID, 
1996_to_1997, 
2000_to_2001,
2005_to_2006,
2010_to_2011, 
Growth_from_2000_to_2010) VALUES(%s,%s,%s,%s,%s,%s)
""", list(df_Economy.values))

mycursor.execute('select * from India_population ')
len(mycursor.fetchall())

#### 2.4 Commit the changes to be updated in the MySQL Server.

cnx.commit()

### 3. India Covid cases by Region.

#### 3.1 Check the data types form DF_India_Covid.

df_covid=pd.read_csv('data/DF_India_Covid.csv')

df_covid.dtypes

df_covid.head()

#### 3.2 Create Table India_Covid_cases:

* Assing the PK (Primary Key) to the column "States_India_ID". 
* Create the columns, 
* Specifying their value types .

sql_3 = """ CREATE TABLE India_Covid_cases(
States_India_ID VARCHAR(50) NOT NULL, 
Confirmed_Cases INT NOT NULL, 
Active_Cases INT NOT NULL, 
Cured INT NOT NULL,
Death INT NOT NULL,
PRIMARY KEY (States_India_ID))

"""

mycursor.execute(sql_3)

mycursor.execute('select * from India_Covid_cases')
mycursor.fetchall()

#### 3.3 Insert the data to the  Table India_Covid_cases:

df_Covid=pd.read_csv('data/DF_India_Covid.csv')

mycursor.executemany("""
INSERT INTO India_Covid_cases(
States_India_ID, 
Confirmed_Cases, 
Active_Cases,
Cured, Death) VALUES(%s,%s,%s,%s,%s)
""", list(df_Covid.values))

mycursor.execute('select * from India_Covid_cases')
len(mycursor.fetchall())

#### 3.4 Commit the changes to be updated in the MySQL Server.

cnx.commit()

### Queries:

#### 1st Querie:

#### Join all tables:

pd.read_sql('''
select *

from India_population

left join India_NSDP_Evolution on India_population.States_India_ID = India_NSDP_Evolution.States_India_ID
left join India_Covid_cases on India_NSDP_Evolution.States_India_ID = India_Covid_cases.States_India_ID
''',cnx).head(15)

#### 2nd Queries :

Convert  Indian Rupee to Euros, **1₹ = 0.0116941€**

### **Do an estimation of the actual population in la india:**

pd.read_sql('''
select 
India_population.States_India_ID,
FORMAT(India_population.year_2011,0) as Population_2011,
CONCAT("€",FORMAT((India_NSDP_Evolution.2010_to_2011 * 0.0116941),2)) as NSDP_2010_to_2011,
FORMAT(India_Covid_cases.Confirmed_Cases,0) as Comfirmed_positive,
FORMAT(India_Covid_cases.Cured,0) as Cured,
FORMAT(India_Covid_cases.Death,0) as Death

from India_population

left join India_NSDP_Evolution on India_population.States_India_ID = India_NSDP_Evolution.States_India_ID
left join India_Covid_cases on India_NSDP_Evolution.States_India_ID = India_Covid_cases.States_India_ID
''',cnx)

* Obtener valores en columnas relevantes(POPULATION ,NSDP_2010TO_2011, CURED, DEATH) MIN, MAX, AVERAGE...

#### 3rd Queries:

Look up and Insert the None Data in Tables, India_NSDP_Evolution and India_Covid_cases:
* Dadra and Nagar Haveli
* Daman and Diu

#### 4th Querie:

Select columns, POPULATION ,NSDP_2010_to_2011, CURED, DEATH and get some conclussions. 

pd.read_sql('''
select India_NSDP_Evolution.States_India_ID,
CONCAT("€",FORMAT((India_NSDP_Evolution.1996_to_1997 * 0.0116941),2)) as NSDP_1996_to_1997,
CONCAT("€",FORMAT((India_NSDP_Evolution.2000_to_2001 * 0.0116941),2)) as NSDP_2000_to_2001,
CONCAT("€",FORMAT((India_NSDP_Evolution.2005_to_2006 * 0.0116941),2)) as NSDP_2005_to_2006,
CONCAT("€",FORMAT((India_NSDP_Evolution.2010_to_2011 * 0.0116941),2)) as NSDP_2010_to_2011,
FORMAT(Cured,0) as Cured,
FORMAT(Death,0) as Death

from India_population

left join India_NSDP_Evolution on India_population.States_India_ID = India_NSDP_Evolution.States_India_ID
left join India_Covid_cases on India_NSDP_Evolution.States_India_ID = India_Covid_cases.States_India_ID
''',cnx)

### VISUALIZATION!!! IMPORTANT

### Finally, close connection:

cnx.close()
